Added the invoke section for the pipeline and moved all socket clients and messages to a new project
