+++ bld_commands/src/monit/mod.rs
@@ -1,5 +1,3 @@
-mod client;
 mod command;
 
-pub use client::*;
 pub use command::*;
