+++ bld_commands/src/run/invoke.rs
@@ -6,8 +6,8 @@ use awc::Client;
 use bld_config::BldConfig;
 use bld_core::logger::Logger;
 use bld_runner::RunnerBuilder;
-use bld_server::requests::RunInfo;
-use bld_server::sockets::ExecClient;
+use bld_sock::messages::RunInfo;
+use bld_sock::clients::ExecClient;
 use bld_utils::request::{self, headers};
 use futures::stream::StreamExt;
 use std::collections::HashMap;
