+++ bld_commands/src/worker/command.rs
@@ -13,8 +13,8 @@ use bld_core::execution::Execution;
 use bld_core::logger::Logger;
 use bld_core::proxies::PipelineFileSystemProxy;
 use bld_runner::RunnerBuilder;
-use bld_supervisor::base::WorkerMessages;
-use bld_supervisor::sockets::WorkerClient;
+use bld_sock::messages::WorkerMessages;
+use bld_sock::clients::WorkerClient;
 use clap::{Arg, ArgAction, ArgMatches, Command};
 use futures::join;
 use futures::stream::StreamExt;
