+++ bld_runner/src/sync/pipeline.rs
@@ -43,6 +43,7 @@ impl Variable {
 pub struct BuildStep {
     pub name: Option<String>,
     pub working_dir: Option<String>,
+    pub invoke: Vec<String>,
     pub call: Vec<String>,
     pub commands: Vec<String>,
 }
@@ -51,12 +52,14 @@ impl BuildStep {
     pub fn new(
         name: Option<String>,
         working_dir: Option<String>,
+        invoke: Vec<String>,
         call: Vec<String>,
         commands: Vec<String>,
     ) -> Self {
         Self {
             name,
             working_dir,
+            invoke,
             call,
             commands,
         }
@@ -168,30 +171,46 @@ impl Pipeline {
     fn steps(yaml: &Yaml) -> Vec<BuildStep> {
         let mut steps = Vec::<BuildStep>::new();
         let working_dir = yaml["working-dir"].as_str().map(|w| w.to_string());
-        if let Some(entries) = &yaml["steps"].as_vec() {
-            for step in entries.iter() {
-                let name = step["name"].as_str().map(|n| n.to_string());
-                let working_dir = step["working-dir"]
-                    .as_str()
-                    .map(|w| w.to_string())
-                    .or_else(|| working_dir.clone());
-                let call = step["call"]
-                    .as_vec()
-                    .unwrap_or(&Vec::<Yaml>::new())
+        yaml["steps"]
+            .as_vec()
+            .map(|steps| {
+                steps
                     .iter()
-                    .map(|c| c.as_str().unwrap_or("").to_string())
-                    .filter(|c| !c.is_empty())
-                    .collect();
-                let commands: Vec<String> = step["exec"]
-                    .as_vec()
-                    .unwrap_or(&Vec::<Yaml>::new())
-                    .iter()
-                    .map(|c| c.as_str().unwrap_or("").to_string())
-                    .filter(|c| !c.is_empty())
-                    .collect();
-                steps.push(BuildStep::new(name, working_dir, call, commands));
-            }
-        }
-        steps
+                    .map(|step| {
+                        let name = step["name"].as_str().map(|n| n.to_string());
+                        let working_dir = step["working-dir"]
+                            .as_str()
+                            .map(|w| w.to_string())
+                            .or_else(|| working_dir.clone());
+
+                        let invoke = step["invoke"]
+                            .as_vec()
+                            .unwrap_or(&Vec::<Yaml>::new())
+                            .iter()
+                            .map(|c| c.as_str().unwrap_or("").to_string())
+                            .filter(|c| !c.is_empty())
+                            .collect();
+
+                        let call = step["call"]
+                            .as_vec()
+                            .unwrap_or(&Vec::<Yaml>::new())
+                            .iter()
+                            .map(|c| c.as_str().unwrap_or("").to_string())
+                            .filter(|c| !c.is_empty())
+                            .collect();
+
+                        let commands: Vec<String> = step["exec"]
+                            .as_vec()
+                            .unwrap_or(&Vec::<Yaml>::new())
+                            .iter()
+                            .map(|c| c.as_str().unwrap_or("").to_string())
+                            .filter(|c| !c.is_empty())
+                            .collect();
+
+                        BuildStep::new(name, working_dir, invoke, call, commands)
+                    })
+                    .collect()
+            })
+            .unwrap_or_else(|| vec![])
     }
 }
