+++ bld_server/src/sockets/mod.rs
@@ -1,11 +1,7 @@
-mod enqueue_client;
 mod exec;
-mod exec_client;
 mod ha;
 mod monit;
 
-pub use enqueue_client::*;
 pub use exec::*;
-pub use exec_client::*;
 pub use ha::*;
 pub use monit::*;
