+++ bld_sock/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "bld_commands"
+name = "bld_sock"
 version = "0.1.0"
 edition = "2021"
 
@@ -8,24 +8,16 @@ edition = "2021"
 [dependencies]
 actix = "0.13.0"
 actix-codec = "0.5.0"
+actix-http = "3.0.4"
 actix-web = { version = "4.0.1", features = ["openssl"] }
-anyhow = "1.0.40"
+actix-web-actors = "4.1.0"
 awc = { version = "3.0.0", features = ["openssl"] }
+anyhow = "1.0.40"
 bld_config = { path = "../bld_config" }
-bld_utils = { path = "../bld_utils" }
-bld_core = { path = "../bld_core" }
-bld_sock = { path = "../bld_sock" }
-bld_runner = { path = "../bld_runner" }
-bld_server = { path = "../bld_server" }
-bld_supervisor = { path = "../bld_supervisor" }
-chrono = "0.4.19"
-clap = "4.0.18"
 futures = "0.3.15"
-oauth2 = "4.0.0"
 serde = "1.0.126"
 serde_derive = "1.0.126"
 serde_json = "1.0.64"
 tokio = { version = "1.15", features = ["full"] }
 tracing = "0.1.36"
-uuid = { version = "0.8.2", features = ["v4"] }
-tabled = "0.9.0"
+openssl = "0.10.42"
