+++ bld_supervisor/Cargo.toml
@@ -16,6 +16,7 @@ anyhow = "1.0.40"
 async-trait = "0.1.50"
 bld_config = { path = "../bld_config" }
 bld_core = { path = "../bld_core" }
+bld_sock = { path = "../bld_sock" }
 diesel = { version = "2.0.2", features = ["sqlite", "r2d2"] }
 futures = "0.3.15"
 serde = "1.0.126"
