+++ bld_supervisor/src/base/mod.rs
@@ -1,5 +1,3 @@
-mod message;
 mod queue;
 
-pub use message::*;
 pub use queue::*;
