+++ bld_supervisor/src/sockets/mod.rs
@@ -1,7 +1,5 @@
 mod server;
 mod worker;
-mod worker_client;
 
 pub use server::*;
 pub use worker::*;
-pub use worker_client::*;
